+++ src/org/processmining/plugins/InductiveMiner/mining/fallthrough/FallThroughFlower.java
@@ -21,8 +21,8 @@ public class FallThroughFlower implements FallThrough {
 		Block loopNode = new AbstractBlock.XorLoop("");
 		loopNode.setProcessTree(tree);
 		MinerMetrics.attachNumberOfTracesRepresented(loopNode, logInfo);
-		MinerMetrics.attachMovesOnLog(loopNode, 0);
-		MinerMetrics.attachMovesOnModelWithoutEpsilonTracesFiltered(loopNode, 0);
+		MinerMetrics.attachMovesOnLog(loopNode, (long) 0);
+		MinerMetrics.attachMovesOnModelWithoutEpsilonTracesFiltered(loopNode, (long) 0);
 		MinerMetrics.attachProducer(loopNode, "fall through: flower model");
 		
 		//body: tau
@@ -30,9 +30,9 @@ public class FallThroughFlower implements FallThrough {
 		body.setProcessTree(tree);
 		loopNode.addChild(body);
 		//count the number of times this tau was used
-		MinerMetrics.attachNumberOfTracesRepresented(body, (int) (logInfo.getNumberOfTraces() + logInfo.getNumberOfEvents()));
-		MinerMetrics.attachMovesOnLog(body, 0);
-		MinerMetrics.attachMovesOnModelWithoutEpsilonTracesFiltered(body, 0);
+		MinerMetrics.attachNumberOfTracesRepresented(body, logInfo.getNumberOfTraces() + logInfo.getNumberOfEvents());
+		MinerMetrics.attachMovesOnLog(body, (long) 0);
+		MinerMetrics.attachMovesOnModelWithoutEpsilonTracesFiltered(body, (long) 0);
 		MinerMetrics.attachProducer(body, "fall through: flower model");
 		
 		//redo: xor/activity
@@ -43,9 +43,9 @@ public class FallThroughFlower implements FallThrough {
 			xorNode = new AbstractBlock.Xor("");
 			xorNode.setProcessTree(tree);
 			loopNode.addChild(xorNode);
-			MinerMetrics.attachNumberOfTracesRepresented(xorNode, (int) logInfo.getNumberOfEvents());
-			MinerMetrics.attachMovesOnLog(xorNode, 0);
-			MinerMetrics.attachMovesOnModelWithoutEpsilonTracesFiltered(xorNode, 0);
+			MinerMetrics.attachNumberOfTracesRepresented(xorNode, logInfo.getNumberOfEvents());
+			MinerMetrics.attachMovesOnLog(xorNode, (long) 0);
+			MinerMetrics.attachMovesOnModelWithoutEpsilonTracesFiltered(xorNode, (long) 0);
 			MinerMetrics.attachProducer(xorNode, "fall through: flower model");
 		}
 		
@@ -55,8 +55,8 @@ public class FallThroughFlower implements FallThrough {
 			xorNode.addChild(child);
 			
 			MinerMetrics.attachNumberOfTracesRepresented(child, logInfo.getActivities().getCardinalityOf(activity));
-			MinerMetrics.attachMovesOnLog(child, 0);
-			MinerMetrics.attachMovesOnModelWithoutEpsilonTracesFiltered(child, 0);
+			MinerMetrics.attachMovesOnLog(child, (long) 0);
+			MinerMetrics.attachMovesOnModelWithoutEpsilonTracesFiltered(child, (long) 0);
 			MinerMetrics.attachProducer(child, "fall through: flower model");
 		}
 		
@@ -64,8 +64,8 @@ public class FallThroughFlower implements FallThrough {
 		tau2.setProcessTree(tree);
 		loopNode.addChild(tau2);
 		MinerMetrics.attachNumberOfTracesRepresented(tau2, logInfo);
-		MinerMetrics.attachMovesOnLog(tau2, 0);
-		MinerMetrics.attachMovesOnModelWithoutEpsilonTracesFiltered(tau2, 0);
+		MinerMetrics.attachMovesOnLog(tau2, (long) 0);
+		MinerMetrics.attachMovesOnModelWithoutEpsilonTracesFiltered(tau2, (long) 0);
 		MinerMetrics.attachProducer(tau2, "fall through: flower model");
 		
 		return loopNode;
